Feat/approve reject delete match cats

+++ internal/match/service/match_service.go
@@ -105,8 +105,46 @@ func (s MatchService) Delete(ctx context.Context, matchID int, userID int) error
 		return msg.Unauthorization("either cat or match request don't belong to user")
 	}
 
+	//Check if match already processed
+	if match.ApprovedAt.Valid || match.RejectedAt.Valid {
+		return msg.BadRequest("matchId is already approved / rejected")
+	}
+
 	// Delete the match
 	err = s.matchRepo.DeleteMatchByMatchId(ctx, matchID)
+	if err != nil {
+		return msg.BadRequest(err.Error())
+	}
+
+	return nil
+}
+
+func (s MatchService) RejectMatchRequest(ctx context.Context, matchParam entity.ProcessMatchRequest, userID int) error {
+
+	//Get Match Info
+	match, err := s.matchRepo.GetMatchByID(ctx, int(matchParam.MatchId))
+	if err != nil {
+		return msg.BadRequest(err.Error())
+	}
+
+	// Fetch the cat information
+	matchCat, err := s.catRepo.GetCatByID(ctx, int(match.IdMatchedCat))
+	if err != nil {
+		return msg.BadRequest(err.Error())
+	}
+
+	// Check if userCatId belongs to the user
+	if matchCat.IdUser != uint32(userID) {
+		return msg.Unauthorization("either cat or match request don't belong to user")
+	}
+
+	//Check if match already processed
+	if match.ApprovedAt.Valid || match.RejectedAt.Valid {
+		return msg.BadRequest("matchId is no longer valid")
+	}
+
+	// Delete the match
+	err = s.matchRepo.RejectByMatchId(ctx, int(matchParam.MatchId))
 	if err != nil {
 		return msg.BadRequest(err.Error())
 	}
